# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NW London Heritage & Planning Engine API
  description: |
    Enterprise-grade API for checking property heritage and planning status in North West London.
    
    ## Overview
    This API provides instant lookup of property planning constraints including:
    - Listed Building status (Grade I, II*, II)
    - Conservation Area boundaries
    - Article 4 Direction restrictions
    
    ## Coverage Area
    - **Boroughs**: Camden, Barnet, Westminster, Haringey, Brent
    - **Postcodes**: NW1-NW11, N2, N6, N10
    
    ## Rate Limiting
    - Default: 60 requests per minute
    - Rate limit headers are included in all responses
    
    ## Authentication
    - Public endpoints (property check, geocoding) do not require authentication
    - Admin endpoints require JWT authentication via NextAuth.js
    
  version: 1.0.0
  contact:
    name: Heritage Checker Support
    email: support@hampstead-heritage.co.uk
    url: https://hampstead-heritage.co.uk
  license:
    name: Proprietary
    url: https://hampstead-heritage.co.uk/terms
servers:
  - url: https://api.hampstead-heritage.co.uk/v1
    description: Production server
  - url: https://staging-api.hampstead-heritage.co.uk/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /generate-report:
    post:
      tags:
        - Property Check
      summary: Generate PDF report
      description: Generates a PDF report for a property check result (server-side)
      operationId: generateReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportGenerationRequest'
      responses:
        '200':
          description: PDF generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ReportGenerationRequest:
      type: object
      required:
        - searchId
        - email
      properties:
        searchId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        includeMap:
          type: boolean
          default: true
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          description: Human-readable error message
        errorCode:
          type: string
          enum:
            - VALIDATION_ERROR
            - GEOCODING_FAILED
            - DATABASE_ERROR
            - RATE_LIMITED
            - NOT_IN_COVERAGE_AREA
            - SERVER_ERROR
          description: Machine-readable error code
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from NextAuth.js authentication
security: []